gtk-ui: implement bread-crumb navigation
The Window widget now supports two 'levels' of view hierarchy, the
window child as top level and arbitrary number of second level views.
It shows the current view as 'bread crumb' buttons in the window
title bar.

The Sync ui currently uses these views:
 * Sync
   * Settings
   * Sync Emergency (not accessible yet)

The graphic design is not final