gtk-ui: Implement sync mode configuration according to new design
Service configuration form now includes sync modes. The full
complexity is not exposed, as the mode is not per source:
all sources in the service share a mode setting, but each have
a enabled-toggle.

So, some configurations cannot be shown in the UI. The UI can be used
to sync with them however, and there will be a warning in the
configuration form in that case.

This commit does not include the use of Moblin toggle buttons,
everything is still implemented with GtkCheckButtons.