DBus server: add 'Session.Restore' (MB#8144)
Restore is an API to let dbus clients to restore to a specified
state from before or after the selected synchronization.

Session.Restore is an asynchronized API and status signal
'done' is to show that an actual restore process is completed.
The real restore operation is running scheduled by dbus server
after the Session.Restore calling.
Session.run is reused to handle 'Sync' and 'Restore'. To
distinguish them, an enum value is selected.

Progress related info has not been available yet.