XMLRPC backend
The XMLRPC backend allows to connect to a webmail or PIM application by
means of XMLRPC, to use the application as data store for syncevolution.

It depends on the xmlrpc-c library.

The unit test part is not finished and left as copied from the file
backend for later completion.

The copyright is waived as documented in the files.