Super datastore consolidation: only expose the super datasource to user
sub datasources referenced by a super datastore will not be visible to user for
sync. The 'sync' and 'URI' properties is not set and will use the value set
in the super datastore.
The 'type' propery will be checked for consistency (values in sub datasoruce and
super datsource should be the same), if failed a warning will be generated and
continue the sync with the type specified by the super datasource.