command line: print status codes for each source
Printing of per-source status codes was avoided, both at the point
where they were encountered as well as in the final sync report.  The
reason for that was that often, the overall sync status is duplicated
for each source.

But for those cases were each source has a different status or no
error at all, printing that piece of information is useful, so this
patch adds it.

Note that now "no status code or error" for a source means that it
synchronized successfully.