sync client: detect unexpected slow sync and abort (MB #2416)
This solution depends on ABORTDATASTORE() in an <alertscript>,
which is supported by libsynthesis only with an additional patch.
ABORTDATASTORE() is done for each source where the sync mode
requested by the server does not match the expected one.

The check is only added when a slow sync would not be acceptable.
Explicit "slow" and "resfresh-from-server" (which is done under
the hood as a slow sync?!) don't need the check. Distinguishing
unexpected "slow" from "refresh-from-client" has to be done
via ALERTCODE(), SLOWSYNC() returns true in both cases.

Note that currently, "no local data" and "first time sync" are
ignored when checking for acceptable slow syncs. This might have
to be added.

In addition, the code in SyncContext::doSync() is notified via the
"delayedabort" session variable that it should abort after processing
the current message and before sending the reply. This is necessary to
get the <alertscript> executed for all sources. Calling ABORTSESSION()
inside the script was tried first but aborted right away, so that an
unexpected slow sync could only be detected for one source.

Abort handling didn't work for aborting before sending a reply:
- SessionStep would return STEPCMD_SENDDATA
- we would start sending
- abort was checked later

A good point to abort a sync is after the engine has processed
a message and before sending out the reply. This patch adds that
check for STEPCMD_SENDDATA and checkForAbort() directly after
the SessionStep() call.

Because the whole suspend/abort handling is somewhat tricky, this
patch adds debug logging around those state changes.

The code which detects the request to abort also analyzes the sitution
and prints some guidance for the user how to recover from the problem:

  [ERROR] Aborting because of unexpected slow sync for source(s): calendar
  [INFO] Doing a slow synchronization may lead to duplicated items or
  lost data when the server merges items incorrectly. Choosing
  a different synchronization mode may be the better alternative.
  Restart synchronization of affected source(s) with one of the
  following sync modes to recover from this problem:
      slow, refresh-from-server, refresh-from-client

  Analyzing the current state:
      syncevolution --status syncevolution_client calendar

  Running with one of the three modes:
      syncevolution --sync [slow|refresh-from-server|refresh-from-client] syncevolution_client calendar

On the server, doing this check in <alertscript> is not sufficient
because the server does the anchor check after executing the
<alertscript> and thus switches to an unexpected slow sync later on.
Need a different solution for server initiated syncs with phones.