command line: explain error codes (MB #2069)
This patch adds a plain text explanation of the Synthesis and SyncML
error codes. It explains whether the error was found locally or
remotely and includes the original status code, because sometimes
this is more helpful than the text.

The ERROR logging uses these explanations and the sync report
formatter.

This patch also adds some more status codes:
- 418 "already exists"
- 22000 "unexpected slow sync"

The latter is in the range reserved for application codes. We'll
use that for MB #2416.