sync client: accept unexpected slow syncs when no local data
When we can be sure that we have no local data in a source,
then the check for an unexpected slow sync can be suppressed
because there is no risk of duplicates or lost data.

The check depends on having done a backup before. If that was
disabled, the check is enabled because there is no other API
to query the content of the local database.

It would be nice if we had the same information about the server,
but that is not available a) without running the sync and b) not
sent by all servers.