TrackingSyncSource: be more quite about getPeerMimeType()
getPeerMimeType() is implemented by the TrackingSyncSource,
so implementors of the class don't have to bother with it.
Removed documentation of the function (which was copied
from the base class) and moved the function to the "private"
section with the other implemented functions (because users
of the TrackingSyncSource API don't need to call it).