DBus server: automatic termination (MB#7711)
Implement automatic termination in the dbus server. A new class
AutoTerm is used to do this kind of checking. If having attached
clients or any connections, never terminate. Else, dbus server
stays alive for a given duration after each method call.

To check idle duration, a source is added in the main context
and will be checked every 5 seconds. The reason why we don't
add a timeout source with the user specified duration is because
that we can't add and remove frequently, which will make the id of
source interger-overflow since it is always increased. However,
the disadvantage is that there might have somewhat inaccuracy.
If the server is busy, the timeout source might not be called in
every 5 seconds. But since the busy server has not to auto-terminate,
the inaccuracy won't effect the correct behavior.

Allow one dbus client attaching or detaching more than one time. Each
attach has to be matched by one detach. Detaching a client which is
not attaching now is unexpected. The current behavior just ignores it.
If a client has gone and it is attaching to the server, the server
always removes all 'Attach' operations for the client by detecting the
client status.

The option "--duration" or "-d" is added in the command line.
It is calculated in seconds not in minutes. The reason is to allow us
testing it with seconds and more flexible. The default duration is 600
seconds. If it's set as "unlimited", the dbus server never terminates.