config names: replace unsafe characters with underscores (MB #8350)
This automatism makes it easier to create configs with the D-Bus API:
let the user choose a config name, check whether it already exists
(must be done in any case!), if not, use the string, otherwise
complain.

Without the automatism, the D-Bus client would have to know which
characters are safe and tell the user.

A drawback of this mechanism is that command line tool users might
accidentally modify a config instead of creating it, because multiple
different config names map to the same normalized string. This is
considered unlikely.

Another drawback is that the D-Bus client cannot tell what the
normalized config name is. But because it can simply use the
un-normalized one to access the new config, this is not such a big
deal.