SmartPtr: allow tracking of non-pointer objects
A GLib guint event handle can be compared to 0 (the invalid handle),
just like a pointer. Previously, eptr didn't support this because the
template was parameterized with a class and then used pointers to it.

This patch changes the base template so that both pointers and
integral types can be tracked. The base template was renamed to
SmartPtr. eptr became a source compatibility class with the previous
"track pointers" semantic.

GLibEvent is a specialization which uses g_source_remove() to free
the event associated with a guint handle.