SoupTransportAgent: event handling
When an instance was deleted while it had a pending message and
active timeout/abort events, those event sources were not removed.
Using the new smart pointer type for GLib event handles avoids that.

According to the documentation, a handle is always > 0, so initializing
it with 0 as GLibEvent is more appropriate than the -1 used previously
in the code here.

There might still be one problem left: what if wait() is left by
throwing an exception? Should the two event sources be cleared in that
case?