LCS: Longest Common Subsequence
The code was originally written for an improved diff to be used in
synccompare. It improves upon the normal LCS by taking into account
that the some subsequences might be "nicer" than others of the same
length, for example because they cross less "boundaries" inside the
sequences.

The reason for adding it now is for use in a fuzzy manufacturer/model
matching.