Logdir: change the logdir and add context info (MB #8350)
The Logdir has no info about context and is not normalized. This
will have problem when the peer name has arbitrary characters.
Normalize the peer name and adding context info in the log dir.
This could help find the corresponding config and look up the 'PeerName'
in the config for sync ui.