documentation + usage: shared configs, 1.0 features (MB #8497)
Mostly replaced <server> with <config> and added new text about the
new config layout. Did some general updating of the documentation to
match the improvements found in 1.0.

Because sharing of properties is not documented, I added code which
prints that as part of "--print-config". Lists of sync properties are
printed beneath the main sync section in a comment, for source
properties once after all sources.

Added --print-configs and --print-peers as alias for --print-servers.