Logdir: use normalized config name instead of <peer>@<context>
The normalized config name for peers in the default context is just
the peer name, without the @default. Using that makes the directory
names shorter and can be done with less code.

The comparison with peerName + @default can be removed, because that
can never happen when @default is not encoded in the name (?).

If these changes are okay, they should be squashed with the previous
patch before merging into master.