Logdir: getServer() might not include context
SyncContext::getServer() = m_server returns the string as passed in
by the user. This may be without the right context, if the peer name
alone is unique enough to select a config (scheduleworld2 instead
of scheduleworld2@other-context).

What we need is the normalized name as used by SyncConfig. Added
SyncConfig::getConfigName() for that and renamed getServer() to
getPeer(), with suitable documentation.