Logdir: return all peers logs for config name without peer name
If the peer name is empty and context is given, we only match
the context and ignore the comparison of peer name for the
changes of encoding context information in the file name.

Currently for the given logdir, we compare the context of log
name and the given config name. This gives the capability to
extinguish same peer names for different contexts.If the behavior
is not to compare it, it's easy to change.