Testing: fix failures and remove dependency for test-dbus.py (MB#9065)
Due to many changes of source code, the behaviors of
some features are also changed. So clean up the test
cases for these features. Another is to remove dependency
of mannual setup for the test environment.
1) New available services are added in the templates.
2) Change the sync status error code from '500' to '10500' for
TestDBusSyncError
3) Remove the dependency on mananually setup this config. Create
the correct config for TestConnection in its own xdg directory.

Currently the main dependency is for class TestSessionAPIsReal.
It depends on a user configuration named 'dbus_unittest', which could be
used to do synchronization successfully. Users have to set up this before
testing the cases of this class.

Another possible dependency is gnome-keyring in your config. In one
case, the system might pop up a dialog to let testers grant the
access rights for passwords in gnome-keyring. For example, if other
applications store a key/password in the gnome-keyring, which is
the same as the keys in the python tests, the above problem could
occur. It is seldome but please avoid it. A possible solution is to
add an option for gnome-keyring in dbus server.

All cases are passed due to these fixes.