preventing slow syncs: optional, off by default (MB #2416)
The new feature of preventing problematic slow syncs has "problems" of
its own:
- the sync-UI still has to be adapted to deal with prevented slow syncs,
  which might not happen in time for 1.0 beta
- we don't have user feedback about this deviation from previous
  behavior
- the nightly testing runs into unexpected slow syncs with Memotoo
  and Mobical, which do a slow sync when they have no data on the
  server

This patch therefore introduces a configuration option
"preventSlowSync" which is off by default. This is also a good place
to explain the feature. Expert users and testers can enable the
feature per peer.

We can change the default value later on after some more testing. It
will be on without users having to change existing configs, because
default values are not set explicitly in configs.

When we enable this by default, the nightly testing must be
reconfigured to run with it disabled for Memotoo and Mobical (probably
best done explicitly inside client-test).