syncclient_sample_config.xml: added "GENDER" and "SIP"
These were found by Ove Kaaven while working on Maemo 5. To be more
consistent with the other chat handles, SIP is stored in SIP_HANDLE
and with SIP_SLOT mapped to X-EVOLUTION-UI-SLOT. The slot might not be
used on Maemo, but perhaps it will be when these changes migrate into
a normal Evolution.