Maemo backend + testing: unique names, always enabled
Added a _maemo suffix to ensure that there is no naming conflict with
the normal Evolution calendar backend. For example, "ical20" in
client-test refers to the traditional Evolution backend, while
"ical20_maemo" is the new Maemo calendar backend. It is unlikely that
both backends will active at the same time, but who knows...

Also enabled the registration of these test unconditionally, as
changed recently in the other backends. Registering the tests creates
no dependency on CPPUnit headers or libs. The compilation of the
client-test program determines whether integration testing is done,
not the compilation of the backend.