session status/error: when setting status, always update error
Jussi mentioned that he was seeing a 200 "okay" status with an error
of "synchronization process died prematurely". That error description
was set earlier for the session, just in case that it dies. Later it
wasn't overwritten.

When a report is written into a node, it should either set a non-zero
error or reset the property. Removing all properties would also work,
but is less consistent with the semantic of << as "append".