ObexTransportAgent: libbluetooth2 vs 3 check (MB #9289)
The signature of sdp_extract_pdu() and sdp_extract_seqtype()
has changed between libbluetooth2 and libbluetooth3. Some
versions also have a _safe variant.

A compile check seems to be necessary to detect this, because the version
of the .pc file was not updated.

This patch only fixes the source code compatibility with
different libbluetooth versions. Binaries depending on libbluetooth2
don't work when only libbluetooth3 is installed.