SAN + forced slow sync: move flag into SyncSource
The main motivation for this change was that access to that flag was
needed inside the SyncSource while experimenting with slow sync
detection.

In the end, the flag was not really needed there. Applying this patch
might make sense anyway, just in case that a similar need arises
again.

Eventually it might also allow to remove the entire
SourceConfigSpecials structure.