slow sync detection: use <datastoreinitscript> instead of <alertscript> (MB #2416)
Using the <datastoreinitscript> has the advantage that a slow sync
chosen by the engine after <alertscript> was called because of an
anchor mismatch is now caught.

This patch also adds a reminder that the dependency on a database dump
should be removed and slightly modifies the check for slow sync. This is
done with an eye towards enabling this code on the server side (not done
yet).