config: adding sources affects peers in the same context (MB #9329)
When adding a source "foo" in peer "xyz" which doesn't exist in peer
"abc" using the same context, the new source appeared as new source in
"abc" (okay, by design because of shared properties) and was enabled
(not okay).

It was meant to be disabled via the "sync = disabled" default value,
except that this default had not been changed when implementing MB

Changed that default now. Because "sync" is set explicitly when
configuring sources, this should have no side effects.