DBus server: return real passwords for GetConfig (MB#9169)
Check password and return the real passwords from gnome keyring
for GetConfig.

The behavior of askPassword in DBusSync will ask password from dbus
clients if they are not found in gnome-keyring. So a new class is
involved between DBusSync and SyncContext. Its behavior of askPassword
is directly to ask password from gnome keyring. If not found, then
return empty string.

Enable a case to test this kind of scenario.