detecting slow sync: use LOCERR_DATASTORE_ABORT (MB #2416)
The ABORTDATASTORE() handling in the Synthesis engine was fixed
so that it aborts the whole session, without looking at further
datastores.

In order to get our desired behavior (continue checking other sources),
a special LOCERR_DATASTORE_ABORT error code was introduced. We use
that in our ABORTDATASTORE(), then translate it back to the SyncEvolution
specific STATUS_UNEXPECTED_SLOW_SYNC = 22000.

The reasons for keeping that code are:
- the sync-UI checks it
- encoding in the sync-UI that LOCERR_DATASTORE_ABORT == slow sync
  would expose an implementation detail that we might want to change
  in the future