LogDir: fixed finding sessions inside non-standard context with escaped chars
When the context name contains characters which need to be escaped in
the logdir name (like "@client-test-1"), the getting all sessions
in that context failed.

The reason is that parseDirName() returns the peerName without unescaping
it. The context name check then compared the unescaped name against
the escaped one and not detecting the match.

IMHO parseDirName() should hide the escaping mechanism and return unescaped
strings, but this is the more intrusive patch and thus I don't change it.
In any case, something like this should be documented for the parameters
of the function.