SourceList: determine most recent backup of source when making next backup (MB #7708)
This is one half of reusing old backup data. SourceList determines the
most recent backup made when syncing with any peer in the same
context. These peers share the same sources, therefore they can also
share backup data.

The "after" dump is checked first, because it is the more recent one.
The "before" dump of the same session is used when creating
the "after" dump.

This patch also prevents creating empty backup dirs when the backup
operation is not supported by the backend. Because all backends supported
this so far, this wasn't a problem.