SyncSourceRevisions: reuse data files from previous backup (MB #7708)
This patch makes database dumps more space-efficient by reusing
existing files with the same content via hard links. "Same content"
is identified via the rather simple Hash() function, which
avoids reading the old files. We should better use a stronger
hash function or check the actual content?!

If nothing has changed, the only space required on disk for another
dump is for the directory structure and the .ini file.

Another advantage is that synccompare can be changed so that it
avoids including identical files in the comparison.

Note that this patch does not "trust" the revision strings. If it did,
reading the file content could have been replaced by lookup via
luid+rev. The advantage of checking and using the item content is
that changes in how the item is formatted show up properly in
comparisons. Improvements in that formatting lead to better backups,
which would not be the case when reusing old files just because the
revision has not changed.