synccompare: don't include identical files in comparison (MB #7708)
When invoked for directories of files, it is easy to determine which
files are identical by looking at the inodes, which are shared when
the two directories were created by recent syncevolution.

Identical files don't have to be included in the diff, which
considerably speeds up the whole comparison.

If identical files happen to have different inodes, the comparison
works as before. The parameter of Normalize() had to be changed so
that accepts strings, because that was necessary for the shared files
case.